+++ apps/pic32mz_ef_curiosity_v2/snake_game/firmware/src/config/default/gfx/legato/memory/legato_fixedheap.c
@@ -243,9 +243,9 @@ leBool leFixedHeap_Contains(leFixedHeap* heap, void* ptr)
 	if(dataPtr < data || dataPtr >= data + (heap->physicalBlockSize * heap->numElements))
 		return LE_FALSE;
 
-	dataPtr -= (uint32_t)data;
+	dataPtr -= (size_t)data;
 	
-	return ((uint32_t)dataPtr % heap->physicalBlockSize) == LE_FIXEDHEAP_HEADER_SIZE;
+	return ((size_t)dataPtr % heap->physicalBlockSize) == LE_FIXEDHEAP_HEADER_SIZE;
 }
 
 leResult leFixedHeap_Validate(struct leFixedHeap* heap)
